[refactor] Separate game answer functions
