+++ kWordle/ViewModel/MainViewModel.swift
@@ -10,7 +10,6 @@ import SwiftUI
 import Firebase
 import GoogleMobileAds
 
-var generator = RandomNumberGeneratorWithSeed(seed: DateToSeed())
 class MainViewModel: ObservableObject {
     @Published var game: Game
     @Published var isInvalidWordWarningPresented: Bool = false
@@ -353,41 +352,4 @@ extension MainViewModel {
     func getLifeCount() {
         lifeCount = life
     }
-}
-
-func todayAnswer() -> String {
-    generator = RandomNumberGeneratorWithSeed(seed: DateToSeed())
-    let wordDictCount = WordDictManager.shared.wordDictFiveJamo.count
-    let randomAnswer = WordDictManager.shared.wordDictFiveJamo[Int(generator.next()) % wordDictCount].jamo
-    var answers: [String] = []
-    answers.append(randomAnswer)
-    UserDefaults.standard.set(answers, forKey: "todayAnswers")
-    return randomAnswer
-}
-
-func randomAnswerGenerator() -> String {
-    let wordDictCount = WordDictManager.shared.wordDictFiveJamo.count
-    var randomNumber = Int(generator.next())
-    var randomAnswer = WordDictManager.shared.wordDictFiveJamo[randomNumber % wordDictCount].jamo
-    var answers = UserDefaults.standard.stringArray(forKey: "todayAnswers") ?? []
-    var i = 0
-    while answers.contains(randomAnswer) {
-        randomNumber = Int(generator.next())
-        randomAnswer = WordDictManager.shared.wordDictFiveJamo[(randomNumber) % wordDictCount].jamo
-        i += 1
-        if i > 1000 {
-            break
-        }
-    }
-    answers.append(randomAnswer)
-    print("Answer: " + randomAnswer)
-    UserDefaults.standard.set(answers, forKey: "todayAnswers")
-    return randomAnswer
-}
-
-func DateToSeed() -> Int {
-    let date = Date()
-    let today = Calendar.current.dateComponents([.year, .month, .day, .hour], from: date)
-    let todaySeed = today.year! * 10000 + today.month! * 100 + today.day!
-    return todaySeed
 }
